refactor: convert workers/repository/finalise to ts
